refac: Refactor Server.GetWorkflows()
